fix: Repair compile/runtime defects and wire real contamination mechanics
Compile and link fixes (found by building every header standalone and
a two-TU link test):
- distributed_experiment.hpp: missing <map>/<tuple>; nested-struct NSDMI
  default-argument error on ExperimentCoordinator::Config
- debug.hpp: missing <execinfo.h> for backtrace, <cstdlib>/<algorithm>/
  <limits>; std::max on atomic without .load()
- json.hpp: missing <cstring>/<cctype>
- llm_interface.hpp: missing experiment_harness.hpp include (broke
  standalone inclusion)
- evaluation_metrics.hpp: method_name() lacked inline -> multiple
  definition at link time across translation units
- versioned_entity.hpp: missing field initializers in ContaminationEvent

Runtime fixes:
- event_bus.hpp: EventBus was 8.4MB (inline 64K-slot MPSC queue) and
  overflowed the stack when constructed as a local; queue moved to heap.
  This crashed the existing test suite on startup.
- ollama_backend.hpp: streaming path called recv() on hardcoded fd 3;
  added SocketConnection::receive_some() and use it
- order_book.hpp: removed dead iterator variables in vwap()
- test_memory_pool.cpp: ObjectPool<int> violated the pool's
  pointer-size static_assert; use int64_t

Experiment pipeline overhaul (fidelity):
- run_experiment now genuinely exercises the research machinery:
  every event updates a VersionedEntity store, contamination is injected
  into real entity state via ContaminationInjector, and mitigation runs
  ContextReliabilityChecker against that store. Baselines skip the
  reliability layer entirely.
- Layered detection: VersionedContext catches stale states via
  supersession/validity; FullVersionedIntegrity additionally catches
  entity-binding and inference-persistence via value-conflict detection
  and quarantines cross-agent claims. Summary/retrieval contamination
  remains undetectable by the entity store (honest limitation).
- Residual propagation: contaminated events still inside the context
  window keep degrading inference, so method differences now emerge
  mechanistically (FullHistory drowns, windows/filters partially
  recover, versioned methods prune at the source).
- TimeFilter window shortened below experiment span (was 60s > 10s
  total, making it identical to FullHistory)
- ExperimentResult gains detected_contaminations; method_name now
  reported from the configured method (FullVersionedIntegrity was
  displayed as VersionedContext)

New tests (all 29 pass):
- test_versioned_entity.cpp: supersession, temporal expiry, conflict
  detection, reliability checker, selective forgetting
- test_contamination.cpp: all injector paths, injection log/stats,
  seed reproducibility, mitigation-beats-baseline hypothesis check,
  detection-only-in-mitigating-methods
- test_json.cpp: primitives, structures, escapes, roundtrip, failure
  paths, real Ollama response shape

+++ tests/test_memory_pool.cpp
@@ -38,11 +38,12 @@ bool test_object_pool_basic() {
 }
 
 bool test_object_pool_reuse() {
-    ObjectPool<int, 4> pool;
+    // Element type must be at least pointer-sized (free-list constraint)
+    ObjectPool<int64_t, 4> pool;
 
     // Allocate
-    int* a = pool.allocate();
-    int* b = pool.allocate();
+    int64_t* a = pool.allocate();
+    int64_t* b = pool.allocate();
     *a = 42;
     *b = 100;
 
@@ -51,8 +52,8 @@ bool test_object_pool_reuse() {
     pool.deallocate(b);
 
     // Reallocate - should reuse
-    int* c = pool.allocate();
-    int* d = pool.allocate();
+    int64_t* c = pool.allocate();
+    int64_t* d = pool.allocate();
 
     if (c != b && c != a) {
         std::cerr << "Expected memory reuse" << std::endl;
